refactor(protocol): flatten the questions layout to one directory
The layer split cost more than it bought. domain/, ports/, and application/
held two, two, and six files behind 206 lines of re-export barrel, and the
three sub-barrels re-exported ~20 symbols no consumer outside the capability
ever imported. Eleven files elsewhere already deep-imported past the barrel,
so the layering was not buying encapsulation either. Same move as intents/ in
b2aac3a and networks/ in 01f8ca6, and the layout chat/, discovery/, and
premises/ already use.

- Lift the nine source files to questions/ and keep one index.ts as the
  capability barrel; the filename already names the layer (question.schema,
  question.input, question.agent, question.presets, question.env,
  question.tools, question.ask.tool, and the two ports).
- Fold QUD_UNDERSPECIFICATION_RULES into question.presets.ts, its only
  consumer, retiring the 16-line question.qud.ts.
- Share the negotiation candidate/provenance field shape and their identical
  counterparty-eligibility refinement. Both superRefine blocks stated the same
  uptake-only rule in different words; the invariant now has one definition.
- Drop the IND-547 "canonical home" headers and their "legacy path is a thin
  compatibility shim" note — those shims were deleted some time ago.
- Rename the five questioner.*.spec.ts files to mirror the file under test.

10 files in one directory, down from 13 across four. The exported surface is
unchanged: index.ts exports the same 44 names. No behavior change.

Validation, rebased onto 80d10e9: tsc --noEmit clean (protocol + api),
protocol build, both architecture checks, 160/160 questions specs, and 39 pass
in the database-backed api questioner adapter specs against local index_test.
The 4 questioner.pooldedup failures in that run reproduce identically on a
pristine origin/dev checkout, so they predate this branch.
