+++ packages/protocol/src/questions/question.persistence.port.ts
@@ -1,13 +1,11 @@
 /**
- * questions/ports/question.persistence.port — question persistence contracts.
+ * questions/question.persistence.port — question persistence contracts.
  *
  * Protocol-level persistence contract for structured questions generated by
  * QuestionerAgent. Implementations live in the backend and are injected into
  * ProtocolDeps.
- *
- * IND-547: canonical question persistence port.
  */
-import type { Question, QuestionMode, QuestionPurpose, QuestionStrategy, QuestionDetection, QuestionActor, QuestionAnswer, UnderspecificationType } from "../domain/question.schema.js";
+import type { Question, QuestionMode, QuestionPurpose, QuestionStrategy, QuestionDetection, QuestionActor, QuestionAnswer, UnderspecificationType } from "./question.schema.js";
 
 /** Shape accepted by `persist()` — everything needed to insert a question row. */
 export interface PersistableQuestion {
