+++ packages/protocol/src/questions/tests/question.ask.tool.spec.ts
@@ -3,11 +3,11 @@ import { config } from "dotenv";
 config({ path: '.env.test', override: true });
 
 import { describe, it, expect, afterEach } from "bun:test";
-import { createAskUserQuestionTools, setQuestionerAgentForTesting } from "../../questions/application/question.ask.tool.js";
-import type { QuestionerAgent } from "../../questions/application/question.agent.js";
+import { createAskUserQuestionTools, setQuestionerAgentForTesting } from "../question.ask.tool.js";
+import type { QuestionerAgent } from "../question.agent.js";
 import type { ResolvedToolContext } from "../../shared/agent/tool.helpers.js";
-import type { PersistableQuestion, PersistedQuestion, ChatQuestionAnswerOutcome, ChatQuestionsHost } from "../../questions/ports/question.persistence.port.js";
-import type { Question } from "../../questions/domain/question.schema.js";
+import type { PersistableQuestion, PersistedQuestion, ChatQuestionAnswerOutcome, ChatQuestionsHost } from "../question.persistence.port.js";
+import type { Question } from "../question.schema.js";
 import { requestContext, type TraceEmitter } from "../../shared/observability/request-context.js";
 
 const userId = '00000000-0000-4000-8000-000000000001';
