+++ packages/protocol/src/questions/tests/question.env.spec.ts
@@ -1,7 +1,7 @@
 /** Unit tests for the centralized questioner env accessors (hierarchy + parsing). */
 import { afterEach, describe, expect, it } from "bun:test";
 
-import { isQuestionerEnabled, isUptakeGuardEnabled, uptakeAuthorityThreshold, chatQuestionWaitTimeoutMs, CHAT_QUESTION_WAIT_TIMEOUT_MS_DEFAULT, UPTAKE_AUTHORITY_THRESHOLD_DEFAULT } from "../../questions/application/question.env.js";
+import { isQuestionerEnabled, isUptakeGuardEnabled, uptakeAuthorityThreshold, chatQuestionWaitTimeoutMs, CHAT_QUESTION_WAIT_TIMEOUT_MS_DEFAULT, UPTAKE_AUTHORITY_THRESHOLD_DEFAULT } from "../question.env.js";
 
 const VARS = [
   "QUESTIONER_ENABLED",
