+++ packages/protocol/src/questions/tests/question.input.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'bun:test';
 
-import { isValidQuestionerInputContract, type QuestionerInput } from '../../questions/application/question.input.js';
+import { isValidQuestionerInputContract, type QuestionerInput } from '../question.input.js';
 
 const base = {
   userId: 'user-1',
