+++ services/api/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@indexnetwork/api",
-  "version": "0.95.0",
+  "version": "0.95.1",
   "license": "MIT",
   "private": true,
   "scripts": {
