+++ services/api/tests/mcp.answer-pending-question.composition.test.ts
@@ -3,7 +3,7 @@ import { readFileSync } from 'node:fs';
 
 const mcp = readFileSync(new URL('../src/controllers/mcp.controller.ts', import.meta.url), 'utf8');
 const factory = readFileSync(new URL('../../../packages/protocol/src/shared/agent/tool.factory.ts', import.meta.url), 'utf8');
-const tools = readFileSync(new URL('../../../packages/protocol/src/questions/application/question.tools.ts', import.meta.url), 'utf8');
+const tools = readFileSync(new URL('../../../packages/protocol/src/questions/question.tools.ts', import.meta.url), 'utf8');
 const direct = readFileSync(new URL('../src/services/tool.service.ts', import.meta.url), 'utf8');
 
 describe('MCP/direct answer_pending_question composition', () => {
